training: fp16 end-to-end (float16 datasets + AMP); quantize windows to 1e-2
- store windows as float16 in build/smooth/swt/cskv; datasets emit half tensors
- train/evaluate/sfoa/smoke: enable fp16 mixed precision via Accelerate, wrap forwards in autocast, guard CPU path
- force deterministic algorithms in sfoa/evaluate
- deploy: round predicted/current loads to 2 decimals

+++ core/dataset.py
@@ -95,8 +95,8 @@ class ShardedWindowsDataset(Dataset):
         if x_arr.ndim == 1:
             x_arr = x_arr[:, None]
 
-        x_t = torch.from_numpy(x_arr).float()
-        y_t = torch.from_numpy(y_arr).float()
+        x_t = torch.from_numpy(x_arr).half()
+        y_t = torch.from_numpy(y_arr).half()
         sid_t = torch.tensor(sid_val, dtype=torch.long)
 
         return x_t, y_t, sid_t
